pyproject.toml
[project]
name = "mar-did"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "channels>=4.3.1",
    "concurrent-log-handler>=0.9.28",
    "coverage>=7.10.5",
    "crispy-bootstrap5>=2025.6",
    "daphne>=4.2.1",
    "django>=5.2.5",
    "django-azure-auth>=2.3.0",
    "django-crispy-forms>=2.4",
    "django-environ>=0.12.0",
    "django-pandas>=0.6.7",
    "environ>=1.0",
    "factory-boy>=3.3.3",
    "faker>=37.5.3",
    "oracledb>=3.3.0",
    "pandas>=2.3.1",
    "psycopg2-binary>=2.9.10",
]
